--- a/fbpcs/private_computation/service/pcf2_shard_combiner_stage_service.py
+++ b/fbpcs/private_computation/service/pcf2_shard_combiner_stage_service.py
@@ -19,6 +19,7 @@ from fbpcs.private_computation.entity.pcs_feature import PCSFeature
 from fbpcs.private_computation.entity.private_computation_instance import (
     PrivateComputationInstance,
     PrivateComputationInstanceStatus,
+    PrivateComputationRole,
 )
 from fbpcs.private_computation.entity.product_config import (
     AttributionConfig,
@@ -149,6 +150,9 @@ class ShardCombinerStageService(PrivateComputationStageService):
             for arg in game_args:
                 arg["visibility"] = result_visibility
 
+        should_wait_spin_up: bool = (
+            pc_instance.infra_config.role is PrivateComputationRole.PARTNER
+        )
         mpc_instance = await create_and_start_mpc_instance(
             mpc_svc=self._mpc_service,
             instance_id=pc_instance.infra_config.instance_id
@@ -164,6 +168,7 @@ class ShardCombinerStageService(PrivateComputationStageService):
             game_args=game_args,
             container_timeout=self._container_timeout,
             repository_path=binary_config.repository_path,
+            wait_for_containers_to_start_up=should_wait_spin_up,
         )
 
         logging.info("MPC instance started running for PCF2.0 Shard Combiner.")